refactor: consolidate motion variants
