+++ src/lib/variants.ts
@@ -1,36 +1,59 @@
-export const containerVariants = {
+import { Variants } from 'framer-motion';
+
+export const containerVariants: Variants = {
   hidden: { opacity: 0 },
   visible: {
     opacity: 1,
     transition: {
-      staggerChildren: 0.2,
+      staggerChildren: 0.05,
       delayChildren: 0.1
     }
   }
 };
 
-export const itemVariants = {
-  hidden: {
-    opacity: 0,
-    y: 30,
-    scale: 0.95
-  },
+export const itemVariants: Variants = {
+  hidden: { opacity: 0, y: 20 },
   visible: {
     opacity: 1,
     y: 0,
-    scale: 1
+    transition: {
+      duration: 0.3,
+      ease: 'easeOut'
+    }
   }
 };
 
-export const cardVariants = {
-  hidden: {
-    opacity: 0,
-    y: 40,
-    scale: 0.95
-  },
+export const cardVariants: Variants = {
+  hidden: { opacity: 0, y: 20 },
   visible: {
     opacity: 1,
     y: 0,
-    scale: 1
+    transition: {
+      duration: 0.4,
+      ease: 'easeOut'
+    }
+  }
+};
+
+export const slideInVariants: Variants = {
+  hidden: { opacity: 0, x: 12 },
+  visible: {
+    opacity: 1,
+    x: 0,
+    transition: {
+      duration: 0.4,
+      ease: 'easeOut'
+    }
+  }
+};
+
+export const headerVariants: Variants = {
+  hidden: { opacity: 0, y: -20 },
+  visible: {
+    opacity: 1,
+    y: 0,
+    transition: {
+      duration: 0.3
+    }
   }
 };
